starting to move Nano S UI code out of main
